+++ unime/src/lib/static/svg/logo/UniMeLogo.svelte
